feat: added Notes table to the database and functionality to retrieve it.
The database has a new table called notes that will represent the users notes. When logging into the app and going to the dashboard, notes will now be retrieved based on the current user logged in.

Currently, notes is being displayed in raw json format into the frontend. But it works and the frontend now has all the data it needs to display notes on the screen.